Puncs are close to matching now..

+++ poincare_clean/fieldline_tracing/RK4.h
@@ -6,13 +6,13 @@
 #include <cmath>
 #include <iostream>
 
-std::pair<float, float> RK4_FLT1(
-    float xStart, float yStart, float zStart,
-    const std::vector<std::vector<std::vector<float>>>& dxdy,
-    const std::vector<std::vector<std::vector<float>>>& dzdy,
-    std::vector<float>& xarray,
-    std::vector<float>& zarray,
+std::pair<double, double> RK4_FLT1(
+    double xStart, double yStart, double zStart,
+    const std::vector<std::vector<std::vector<double>>>& dxdy,
+    const std::vector<std::vector<std::vector<double>>>& dzdy,
+    std::vector<double>& xarray,
+    std::vector<double>& zarray,
     int region,
-    const std::vector<std::vector<float>>& dxdy_pm1,
-    const std::vector<std::vector<float>>& dzdy_pm1,
+    const std::vector<std::vector<double>>& dxdy_pm1,
+    const std::vector<std::vector<double>>& dzdy_pm1,
     int direction, int nypf1, int nypf2);
